+++ vec3.h
@@ -6,7 +6,7 @@
 
 class Vector3 {
 public:
-    double x, y, z;
+    float x, y, z;
 
     // Default Constructor
     Vector3() {
@@ -16,7 +16,7 @@ public:
     }
 
     // Parameterized Constructor
-    Vector3(double x, double y, double z) {
+    Vector3(float x, float y, float z) {
         this->x = x;
         this->y = y;
         this->z = z;
@@ -41,12 +41,12 @@ inline Vector3 operator*(const Vector3& v1, const Vector3& v2) {
     return Vector3(v1.x * v2.x, v1.y * v2.y, v1.z * v2.z);
 }
 
-inline Vector3 operator*(const Vector3& v, double scalar) {
+inline Vector3 operator*(const Vector3& v, float scalar) {
     return Vector3(v.x * scalar, v.y * scalar, v.z * scalar);
 }
 
 // Allows scalar * vector order multiplication (e.g., 2.5 * velocity)
-inline Vector3 operator*(double scalar, const Vector3& v) {
+inline Vector3 operator*(float scalar, const Vector3& v) {
     return Vector3(v.x * scalar, v.y * scalar, v.z * scalar);
 }
 
@@ -54,7 +54,7 @@ inline Vector3 operator/(const Vector3& v1, const Vector3& v2) {
     return Vector3(v1.x / v2.x, v1.y / v2.y, v1.z / v2.z);
 }
 
-inline Vector3 operator/(const Vector3& v1, double scalar) {
+inline Vector3 operator/(const Vector3& v1, float scalar) {
     return Vector3(v1.x / scalar, v1.y / scalar, v1.z / scalar);
 }
 
@@ -62,7 +62,7 @@ inline Vector3 neg(const Vector3& v) {
     return Vector3(-v.x, -v.y, -v.z);
 }
 
-inline double dot(const Vector3& v1, const Vector3& v2) {
+inline float dot(const Vector3& v1, const Vector3& v2) {
     return v1.x * v2.x + v1.y * v2.y + v1.z * v2.z;
 }
